update with non-admin routes
