+++ ui/user/src/routes/inventory/AllowlistServerDialog.svelte
