+++ ui/user/src/routes/inventory/Configuration.svelte
