+++ ui/user/src/routes/inventory/ConfigurationDetails.svelte
