+++ ui/user/src/routes/inventory/DeviceClients.svelte
