+++ ui/user/src/routes/inventory/DeviceMcpServers.svelte
