+++ ui/user/src/routes/inventory/DeviceSkills.svelte
