+++ ui/user/src/routes/inventory/EnforcementSettings.svelte
