+++ ui/user/src/routes/inventory/EnrollmentConfigDownload.svelte
