+++ ui/user/src/routes/inventory/GettingStarted.svelte
