+++ ui/user/src/routes/inventory/OverviewView.svelte
@@ -1,12 +1,10 @@
 <script lang="ts">
 	import { resolve } from '$app/paths';
 	import { page } from '$app/state';
-	import TabLayout from '$lib/components/TabLayout.svelte';
 	import AuditLogCalendar from '$lib/components/admin/audit-logs/AuditLogCalendar.svelte';
 	import DeviceScanDonutCard from '$lib/components/admin/device-scan/DeviceScanDonutCard.svelte';
 	import DeviceScanTimelineCard from '$lib/components/admin/device-scan/DeviceScanTimelineCard.svelte';
 	import { buildDeviceScanTopBuckets } from '$lib/components/admin/device-scan/deviceScanTopBuckets';
-	import Devices from '$lib/components/admin/devices/Devices.svelte';
 	import {
 		AdminService,
 		type DeviceClientStat,
@@ -16,10 +14,6 @@
 	} from '$lib/services';
 	import { replaceState } from '$lib/url';
 	import { openUrl } from '$lib/utils';
-	import Configuration from './Configuration.svelte';
-	import DeviceClients from './DeviceClients.svelte';
-	import DeviceMcpServers from './DeviceMcpServers.svelte';
-	import DeviceSkills from './DeviceSkills.svelte';
 	import { DEFAULT_WINDOW_MS } from './constants';
 	import {
 		ChevronRight,
@@ -30,30 +24,34 @@
 		Server,
 		Users
 	} from '@lucide/svelte';
-	import { untrack } from 'svelte';
+	import { onMount, untrack } from 'svelte';
 
-	let { data } = $props();
+	interface Props {
+		stats?: DeviceScanStats | null;
+		range?: { start: string; end: string };
+	}
 
-	type View =
-		| 'configuration'
-		| 'overview'
-		| 'devices'
-		| 'device-clients'
-		| 'device-mcp-servers'
-		| 'device-skills';
-	let stats = $state<DeviceScanStats | null>(untrack(() => data?.stats ?? null));
-	let range = $state<{ start: string; end: string }>(
-		untrack(
-			() =>
-				data?.range ?? {
-					start: new Date(Date.now() - DEFAULT_WINDOW_MS).toISOString(),
-					end: new Date().toISOString()
-				}
-		)
-	);
+	let { stats: initialStats = null, range: initialRange }: Props = $props();
+
+	function rangeFromUrl() {
+		return {
+			start:
+				page.url.searchParams.get('start') ??
+				initialRange?.start ??
+				new Date(Date.now() - DEFAULT_WINDOW_MS).toISOString(),
+			end: page.url.searchParams.get('end') ?? initialRange?.end ?? new Date().toISOString()
+		};
+	}
+
+	let stats = $state<DeviceScanStats | null>(untrack(() => initialStats ?? null));
+	let range = $state<{ start: string; end: string }>(untrack(() => rangeFromUrl()));
 	let loading = $state(false);
 
-	const defaultView: View = untrack(() => (data.configuration ? 'overview' : 'configuration'));
+	onMount(() => {
+		if (range.start !== initialRange?.start || range.end !== initialRange?.end) {
+			reload();
+		}
+	});
 
 	let clientBuckets = $derived(
 		buildDeviceScanTopBuckets<DeviceClientStat>(
@@ -179,49 +177,59 @@
 	]);
 </script>
 
-<svelte:head>
-	<title>Obot | Inventory</title>
-</svelte:head>
-
-<TabLayout
-	title="Inventory"
-	{defaultView}
-	classes={{ childrenContainer: 'max-w-none' }}
-	views={[
-		{ label: 'Configuration', value: 'configuration', content: configuration },
-		{ label: 'Overview', value: 'overview', content: overview },
-		{ label: 'Devices', value: 'devices', content: devices },
-		{ label: 'Device Clients', value: 'device-clients', content: deviceClients },
-		{ label: 'Device MCP Servers', value: 'device-mcp-servers', content: deviceMcpServers },
-		{ label: 'Device Skills', value: 'device-skills', content: deviceSkills }
-	]}
-/>
-
-{#snippet configuration()}
-	<Configuration
-		configuration={data.configuration}
-		enrollmentKeys={data.enrollmentKeys}
-		assetSource={data.assetSource}
-		assets={data.assets}
-		assetLoadError={data.assetLoadError}
+<div class="flex flex-wrap items-center gap-2">
+	<AuditLogCalendar
+		start={new Date(range.start)}
+		end={new Date(range.end)}
+		onChange={onRangeChange}
+		disabled={loading}
 	/>
-{/snippet}
+</div>
 
-{#snippet devices()}
-	<Devices />
-{/snippet}
+{#if !stats || stats.deviceCount === 0}
+	<div class="mx-auto mt-12 flex w-md flex-col items-center gap-4 text-center">
+		<ScanLine class="text-muted-content size-24 opacity-50" />
+		<h4 class="text-muted-content text-lg font-semibold">No device scans in this window</h4>
+		<p class="text-muted-content text-sm font-light">
+			Adjust the date range or run <code class="font-mono">obot scan</code> from a managed device.
+		</p>
+	</div>
+{:else}
+	<div
+		class="paper dark:divide-base-400 divide-base-300 grid grid-cols-2 divide-x sm:grid-cols-3 lg:grid-cols-5"
+	>
+		{#each tiles as tile (tile.key)}
+			{@render statCell(tile)}
+		{/each}
+	</div>
 
-{#snippet deviceClients()}
-	<DeviceClients />
-{/snippet}
-
-{#snippet deviceMcpServers()}
-	<DeviceMcpServers />
-{/snippet}
-
-{#snippet deviceSkills()}
-	<DeviceSkills />
-{/snippet}
+	<div class="grid grid-cols-1 gap-4 lg:grid-cols-2">
+		<DeviceScanDonutCard
+			title="Clients"
+			buckets={clientBuckets}
+			totalGroups={totalClientGroups}
+			emptyMsg="No clients observed yet."
+		/>
+		<DeviceScanDonutCard
+			title="Top MCPs"
+			buckets={mcpBuckets}
+			totalGroups={totalMcpGroups}
+			emptyMsg="No MCP servers observed yet."
+		/>
+		<DeviceScanDonutCard
+			title="Top Skills"
+			buckets={skillBuckets}
+			totalGroups={totalSkillGroups}
+			emptyMsg="No skills observed yet."
+		/>
+		<DeviceScanTimelineCard
+			rangeStart={range.start}
+			rangeEnd={range.end}
+			{timelineRows}
+			totalSubmissions={totalScansInWindow}
+		/>
+	</div>
+{/if}
 
 {#snippet statCell(tile: StatTile)}
 	{#if tile.seeMore === 'devices'}
@@ -274,60 +282,4 @@
 		<span class="text-2xl font-semibold tabular-nums">{tile.value}</span>
 	</div>
 	<tile.icon class="text-primary size-7 shrink-0" />
-{/snippet}
-
-{#snippet overview()}
-	<div class="flex flex-wrap items-center gap-2">
-		<AuditLogCalendar
-			start={new Date(range.start)}
-			end={new Date(range.end)}
-			onChange={onRangeChange}
-			disabled={loading}
-		/>
-	</div>
-
-	{#if !stats || stats.deviceCount === 0}
-		<div class="mx-auto mt-12 flex w-md flex-col items-center gap-4 text-center">
-			<ScanLine class="text-muted-content size-24 opacity-50" />
-			<h4 class="text-muted-content text-lg font-semibold">No device scans in this window</h4>
-			<p class="text-muted-content text-sm font-light">
-				Adjust the date range or run <code class="font-mono">obot scan</code> from a managed device.
-			</p>
-		</div>
-	{:else}
-		<div
-			class="paper dark:divide-base-400 divide-base-300 grid grid-cols-2 divide-x sm:grid-cols-3 lg:grid-cols-5"
-		>
-			{#each tiles as tile (tile.key)}
-				{@render statCell(tile)}
-			{/each}
-		</div>
-
-		<div class="grid grid-cols-1 gap-4 lg:grid-cols-2">
-			<DeviceScanDonutCard
-				title="Clients"
-				buckets={clientBuckets}
-				totalGroups={totalClientGroups}
-				emptyMsg="No clients observed yet."
-			/>
-			<DeviceScanDonutCard
-				title="Top MCPs"
-				buckets={mcpBuckets}
-				totalGroups={totalMcpGroups}
-				emptyMsg="No MCP servers observed yet."
-			/>
-			<DeviceScanDonutCard
-				title="Top Skills"
-				buckets={skillBuckets}
-				totalGroups={totalSkillGroups}
-				emptyMsg="No skills observed yet."
-			/>
-			<DeviceScanTimelineCard
-				rangeStart={range.start}
-				rangeEnd={range.end}
-				{timelineRows}
-				totalSubmissions={totalScansInWindow}
-			/>
-		</div>
-	{/if}
 {/snippet}
