+++ ui/user/src/routes/inventory/[device_id]/+page.svelte
