+++ ui/user/src/routes/inventory/[device_id]/scans/[scan_id]/+layout.ts
