+++ ui/user/src/routes/inventory/[device_id]/scans/[scan_id]/+page.svelte
