+++ ui/user/src/routes/inventory/[device_id]/scans/[scan_id]/_shared/files.ts
