+++ ui/user/src/routes/inventory/[device_id]/scans/[scan_id]/plugins/[id]/+page.svelte
