+++ ui/user/src/routes/inventory/clients/[name]/+page.svelte
