+++ ui/user/src/routes/inventory/constants.ts
