+++ ui/user/src/routes/inventory/mcp-servers/+page.ts
