+++ ui/user/src/routes/inventory/mcp-servers/[hash]/+page.svelte
