+++ ui/user/src/routes/inventory/mcp-servers/[hash]/+page.ts
