+++ ui/user/src/routes/inventory/platforms.ts
