+++ ui/user/src/routes/inventory/skills/+page.ts
