+++ ui/user/src/routes/inventory/skills/[name]/+page.svelte
